cmd/fillstruct: add option to select by line number.
The new command line parameter -line can be used to specify the
position of one or more struct literals by providing the corresponding
line number.

If both parameters, -offset and -line, are provided, then the tool
first uses the more specific offset information. If there was no
struct literal found at the given offset, then the line information is
used.

If more than one struct literal was found, code to fill all of them is
generated. The output is a JSON array with the elements in reverse
order of their occurence.

Fixes #3.